Followed Instructor:  Emmanuel Henri Learning npm course
Initial results of trying to run some of the command examples from
the Learning npm course on LinkedIn Learning.

Probably should have forked the repo, and then cloned the fork,
but oh, well.  Should it not be possible to set up a fork as
a remote for this repo?  Sigh.